Added Trial of Wars Mission 3
Trial of Wars Mission 3 - Data Mine

"The Orange Star have agreed to help us stabilize the planet. To do so, we will build three large Stabilizer devices that will bring stability to the most dangerous spots on the planet. The construction of each of these devices is underway, but to complete them, we need technology from two more of the so-called "armies" on the planet. Commander, I know I'm asking the ridiculous, but you will need to sieze property from these two armies and gather technology from them. We have given you authorization to use brute force in this battle. Capture and defend each of the sides' factories and keep them intact long enough for us to extract the technology we need."

Difficulty: Medium